Add files via upload
So I made this out of curiosity about what and how pytube works, as I was bored reading documentation I searched for examples and sample codes for pytube and came up with a project named youtube video downloader, although I don't know much about pytube I decided to code and learn the code and understand its working. As my project was completed I was able to download a video but it was not in high-resolution quality maybe after some days of studying pytube I may edit some lines of code here. This is awesome project for beginners and also as a college mini project it will be an awesome project to do and will surely impress your teachers